fix(vectorstore): keep vector normalization in float64 precision
normalizeVector accumulated sumSquares in float64 but truncated the norm
to float32 before dividing. A sum near float32's max produces +Inf after
that truncation, zeroing every component and silently giving the chunk a
zero-norm embedding that never scores in cosine similarity.

+++ internal/infrastructure/vectorstore/store.go
@@ -229,10 +229,10 @@ func normalizeVector(vec []float32) []float32 {
 	for _, v := range vec {
 		sumSquares += float64(v) * float64(v)
 	}
-	norm := float32(math.Sqrt(sumSquares))
+	norm := math.Sqrt(sumSquares)
 	out := make([]float32, len(vec))
 	for i, v := range vec {
-		out[i] = v / norm
+		out[i] = float32(float64(v) / norm)
 	}
 	return out
 }
